Adding Tamil language words [see desc]
Hello there! I wanted to contribute to this repository the Tamil language, although some of the words are gtranslate, many of the words are from me, I wrote it.
Although I didn't put every word in here to translate, I did translate a couple of the meals. 
Love the idea of the app! Hope it grows more with more languages, including languages of the Indian Subcontinent, as it differs every 100-200 kms.